requirements.txt
langchain
langchain-community
faiss-cpu
pypdf
langchain-openai
tiktoken
flask
python-dotenv

